Add Docker deployment and document usage
