+++ services/SQLService.js
@@ -1,17 +1,24 @@
 const mysql = require('mysql2/promise');
-const config = require('../config/config');
 const log4js = require('log4js');
+const config = require('../config');
+
+const logger = log4js.getLogger();
+
+const pool = mysql.createPool({
+  ...config.db,
+  waitForConnections: true,
+  connectionLimit: 10,
+  queueLimit: 0,
+});
 
 async function query(sql, params) {
-  const connection = await mysql.createConnection(config.db);
   try {
-    const [results] = await connection.execute(sql, params);
-    connection.end();
+    const [results] = await pool.execute(sql, params);
     return results;
   } catch (err) {
-    connection.end();
-    //logger.error(err); 
-   } 
+    logger.error('Database query error', err);
+    throw err;
+  }
 }
 
 module.exports = {
