+++ services/helperService.js
@@ -1,5 +1,5 @@
 function getOffset(currentPage = 1, listPerPage) {
-  return (currentPage - 1) * [listPerPage];
+  return (currentPage - 1) * listPerPage;
 }
 
 function emptyOrRows(rows) {
